requirements.txt
opencv-python>=4.8.0
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=9.0.0
tkinter-tooltip==2.0.0
scikit-image>=0.19.0
scipy>=1.9.0 